HSEARCH-3970 Remove the TimingSource SPI and move related classes to a common implementation package
1. The SPI makes no sense since it cannot be used (you can't provide
   your own implementation).
2. The timing source is used for more than just search timeouts, so it
   shouldn't be in the search.timeout package.
