+++ backend/lucene/src/main/java/org/hibernate/search/backend/lucene/common/timing/impl/DefaultTimingSource.java
@@ -4,13 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.backend.lucene.search.timeout.impl;
+package org.hibernate.search.backend.lucene.common.timing.impl;
 
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
 import org.hibernate.search.backend.lucene.resources.impl.BackendThreads;
-import org.hibernate.search.backend.lucene.search.timeout.spi.TimingSource;
 
 /**
  * Default implementation for a {@link TimingSource}
